Feat: 불광동 데이터 추가, 리뷰 별점 수정

+++ src/main/java/com/kusitms/forpet/dto/ReviewDto.java
@@ -11,7 +11,7 @@ public class ReviewDto {
     private Long id;
     private String nickName;
     private String profileImageUrl;
-    private int star;
+    private double star;
     private String content;
     private String createDate;
     private String[] imageUrlList;
